# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Infinity Data Explorer API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/units": {
      "get": {
        "tags": [
          "units"
        ],
        "summary": "List Units",
        "description": "List units with optional faction filter.",
        "operationId": "list_units_api_units_get",
        "parameters": [
          {
            "name": "faction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by faction slug",
              "title": "Faction"
            },
            "description": "Filter by faction slug"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnitSummaryResponse"
                  },
                  "title": "Response List Units Api Units Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/units/{slug}": {
      "get": {
        "tags": [
          "units"
        ],
        "summary": "Get Unit",
        "description": "Get full unit details with profiles and loadouts.",
        "operationId": "get_unit_api_units__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "extra": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Extra",
            "default": []
          },
          "extra_display": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extra Display",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ItemRef",
        "description": "A reference to a weapon/skill/equipment with optional modifiers."
      },
      "LoadoutResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "option_id": {
            "type": "integer",
            "title": "Option Id"
          },
          "profile_group_id": {
            "type": "integer",
            "title": "Profile Group Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "swc": {
            "type": "number",
            "title": "Swc"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "equipment": {
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            },
            "type": "array",
            "title": "Equipment",
            "default": []
          },
          "weapons": {
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            },
            "type": "array",
            "title": "Weapons",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "option_id",
          "profile_group_id",
          "name",
          "points",
          "swc"
        ],
        "title": "LoadoutResponse",
        "description": "Equipment option / variant."
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "profile_group_id": {
            "type": "integer",
            "title": "Profile Group Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mov": {
            "type": "string",
            "title": "Mov"
          },
          "cc": {
            "type": "integer",
            "title": "Cc"
          },
          "bs": {
            "type": "integer",
            "title": "Bs"
          },
          "ph": {
            "type": "integer",
            "title": "Ph"
          },
          "wip": {
            "type": "integer",
            "title": "Wip"
          },
          "arm": {
            "type": "integer",
            "title": "Arm"
          },
          "bts": {
            "type": "integer",
            "title": "Bts"
          },
          "wounds": {
            "type": "integer",
            "title": "Wounds"
          },
          "silhouette": {
            "type": "integer",
            "title": "Silhouette"
          },
          "is_structure": {
            "type": "boolean",
            "title": "Is Structure"
          },
          "unit_type": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Type"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "equipment": {
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            },
            "type": "array",
            "title": "Equipment",
            "default": []
          },
          "weapons": {
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            },
            "type": "array",
            "title": "Weapons",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "profile_group_id",
          "name",
          "mov",
          "cc",
          "bs",
          "ph",
          "wip",
          "arm",
          "bts",
          "wounds",
          "silhouette",
          "is_structure"
        ],
        "title": "ProfileResponse",
        "description": "Unit stat line."
      },
...
      "UnitDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "isc": {
            "type": "string",
            "title": "Isc"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "factions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Factions"
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/ProfileResponse"
            },
            "type": "array",
            "title": "Profiles"
          },
          "loadouts": {
            "items": {
              "$ref": "#/components/schemas/LoadoutResponse"
            },
            "type": "array",
            "title": "Loadouts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "isc",
          "name",
          "slug",
          "factions",
          "profiles",
          "loadouts"
        ],
        "title": "UnitDetailResponse",
        "description": "Full unit detail with profiles and loadouts."
      },
      "UnitSummaryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "isc": {
            "type": "string",
            "title": "Isc"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "factions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Factions"
          },
          "points_min": {
            "type": "integer",
            "title": "Points Min"
          },
          "points_max": {
            "type": "integer",
            "title": "Points Max"
          }
        },
        "type": "object",
        "required": [
          "id",
          "isc",
          "name",
          "slug",
          "factions",
          "points_min",
          "points_max"
        ],
        "title": "UnitSummaryResponse",
        "description": "Lightweight unit info for list / search views."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}